This is an enormous commit, and it isn't 100% stable. I just had to commit because too much has changed. The largest changes by far are the incorporation of Laughter into a luckBonus(). I probably modded more than 100 places by adding that in. In most cases I fudged the odds so average-Laughter characters get same results. Seriously, I can't overstate the hoofprint of Laughter's luck bonus: it blocks bad things, improves loot drops, enchant chances, good item types, recovery, etc. etc. It will take a ton of testing to detect any mistakes. Completely redesigned awareness. It was a fucked up mess on crack before. Now it's elegant and simple and works the same with simpler calculations. Generosity is now a big part of awareness. Generosity now affects how fast you heal (mending). Generosity now speeds up pseudo-identification. Haste now includes Laughter as a factor. Will probably need to boost Haste of lower-level monsters to compensate. Speed now includes Pegasus (when unencumbered) and Loyalty as factors. Kindness is now the largest part of Evasion/Defense. Kindness has been rolled into Stealth as well. Magic makes Scrolls of Psionic Blast much more powerful. Magic makes Wands slightly stronger. Magic makes it less likely for glyphs and enchantments to be erased accidentally. (TODO: Maybe change this to give the player a yes/no choice?) Accuracy now includes Loyalty as a factor. This is true in ALL cases, which needed to happen because Loyalty isn't strong enough just with Speed boost and Bows. Tome is now a permanent item: it can't be stolen, dropped, thrown, or generated if you have one or if you have a subclass already. Codexes are being totally redone, with rarity levels (for pricing) and all the secret game information in them. Let's pull most of the tutorial crap out of the level signs. TODO: Prevent Item upgrades at attr = 0. Add magic backfire for rings and wands. Add Magic DR. Add ways to improve Lau, Gen, Kin, Mag. Modify difficulty settings. FIXME: Accuracy for earth ponies is too high, could be any of the new calculations being done. WEAPON_STATUE sprites crash the system now, might be old bug though.

+++ src/main/java/com/watabou/pixeldungeon/actors/mobs/Tengu.java
@@ -28,9 +28,10 @@ import com.watabou.pixeldungeon.actors.Actor;
 import com.watabou.pixeldungeon.actors.Char;
 import com.watabou.pixeldungeon.actors.blobs.ToxicGas;
 import com.watabou.pixeldungeon.actors.buffs.Poison;
+import com.watabou.pixeldungeon.actors.hero.HeroSubClass;
 import com.watabou.pixeldungeon.effects.CellEmitter;
 import com.watabou.pixeldungeon.effects.Speck;
-import com.watabou.pixeldungeon.items.TomeOfMastery;
+import com.watabou.pixeldungeon.items.utility.TomeOfMastery;
 import com.watabou.pixeldungeon.items.keys.SkeletonKey;
 import com.watabou.pixeldungeon.items.scrolls.ScrollOfMagicMapping;
 import com.watabou.pixeldungeon.levels.Terrain;
@@ -74,7 +75,12 @@ public class Tengu extends Boss {
 	
 	@Override
 	public void die( Object cause ) {
-		Dungeon.level.drop( new TomeOfMastery(), getPos() ).sprite.drop();
+
+		// Reset so Tengu only drops tome if you didn't have it at start.
+		// It should be impossible to lose the tome from your inventory now -- can't drop, throw, be stolen.
+		if (Dungeon.hero.subClass == HeroSubClass.NONE && !Dungeon.hero.belongings.backpack.contains(new TomeOfMastery()) ) {
+			Dungeon.level.drop(new TomeOfMastery(), getPos()).sprite.drop();
+		}
 
 		GameScene.bossSlain();
 		Dungeon.level.drop( new SkeletonKey(), getPos() ).sprite.drop();
